+++ tests/testthat/test-utils_callr.R
@@ -1,0 +1,5 @@
+tar_test("callr_args_default", {
+  skip_on_cran()
+  expect_message(out <- callr_args_default(callr::r))
+  expect_equal(out, tar_callr_args_default(callr::r))
+})