+++ NEWS.md
@@ -1,3 +1,7 @@
+# targets 0.13.1
+
+* Fix reverse dependency checks.
+
 # targets 0.13.0
 
 ## Bug fixes